requirements.txt
requests
python-dotenv
jsonschema
collections-extended
matplotlib
numpy
plotly
pandas
seaborn
langchain-core
langchain-groq
flask
matplotlib
numpy
plotly
pandas
seaborn

